pyproject.toml
[project]
name = "bussines-case-jota"
version = "0.1.0"
description = "Bussines case for Jota's Dev Backend vacancy"
authors = [
    {name = "Edimar de Sá",email = "edimar.sa@efscode.com.br"}
]
license = {text = "Unlicense"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.2,<6.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "djangorestframework-simplejwt[crypto] (>=5.5.0,<6.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "pika (>=1.3.2,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
mypy = "^1.15.0"
faker = "^37.1.0"

